refactoring to put popup in different component
